% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenRepGrid.gui-package.R
\name{OpenRepGrid.app}
\alias{OpenRepGrid.app}
\title{Run OpenRepGrid.app}
\usage{
OpenRepGrid.app(display.mode = "auto", launch.browser = TRUE)
}
\arguments{
\item{display.mode}{\code{auto} by default, can also be \code{showcase}.
See \link[shiny]{runApp}.}

\item{launch.browser}{Boolean, set \code{TRUE} to open the app in the browser.
See \link[shiny]{runApp}.}
}
\description{
This function starts the OpenRepGrid GUI app.
}
\examples{
\dontrun{
OpenRepGrid.app()
}

}
