% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenRepGrid.gui-package.R
\docType{package}
\name{OpenRepGrid.gui-package}
\alias{OpenRepGrid.gui-package}
\title{\pkg{OpenRepGrid.gui} - Graphical User Interface to Analyze Repertory Grid Data}
\description{
\pkg{OpenRepGrid.gui} - Graphical User Interface to Analyze Repertory Grid Data
}
\references{
\itemize{
 \item Green, B. (2004). Personal construct psychology and content analysis. Personal Construct Theory & Practice, 1(3), 82-91.
 \item Jankowicz, D. (2004). The easy guide to repertory grids. Chichester, England: John Wiley & Sons.
 }
}
\keyword{package}
\keyword{repgrid}
